Put changes to IPC queue instead of file.
Using a file and filewatcher for interprocess communication seemed to often
cause multiple calls to the watch handler for a given input. Instead, we use
a modified Hotqueue, which builds a durable queue on top of Redis.

On the Clojure side, we're using https://github.com/threatgrid/redismq, which
pops from the right end of the queue (Redis list) so we use a modified version
of Hotqueue that pushes onto the left:
https://github.com/spellman/hotqueue/tree/put-with-lpush